ci: run on mac and windows as well (#78)
Some of the script and server functionality doesn't work from within the
sbt-run scalatests on windows.
With the artifacts of `sbt stage` everything seems to work for windows
though.
I couldn't find a setup that works for both axes and still works for
linux...

This is getting complicated. For now I disabled select tests when run on
windows, so this doesn't hold up
downstream changes any longer. To cover for windows testing I'll add an
integrationtest suite in the future,
i.e. it'll run `sbt stage` and then interact with the staged artifacts
via bash and powershell.