classpath: deduplicate and choose highest version for each artifact (#114)
Goal: for each organisation/name tuple: filter out anything but the
highest version. This is to ensure we don't have jars with different
versions on the classpath.
Problem: since we don't only get the jars via maven coordinates, but
also inherit from existing classpath entries, all we have is a list of
jars.
The best solution is probably to use separate classloaders for the REPL
itself and the code that's executed within. Similar to what ammonite
does..
In the interim, we use this hack as a semi-best-effort basis: group by
the information in their manifests.

Adds coursier.core.Version as a shaded dependency (incl. import
instructions).