fixup line number reporting and simplify implementation: part 1 (#157)
Part 1 (this PR): predef files. Prior to this we concatenated all predef
files
and handed that over to the REPL/script/embedded driver for
interpretation.
That worked, but the line numbers were wrong, and the implementation
more
complicated than necessary. 

A followup PR will address the issue for scripts as well. That's
somewhat
different because we need to embed the script content in some wrapper
code..