split, fix and cleanup predef and runBefore handling (#176)
Prior to 0.1.92 we concatenated and executed all predef code. That was
very simple but had quite
a few downsides, including incorrect line number reporting and some edge
cases when handling complex
compilation units, e.g. ordering of imports etc. 

https://github.com/mpollmeier/scala-repl-pp/pull/157 changed this
behaviour to instead compile
the given files and add the results to the classpath. At the same time
we unintendedly broke
the ability to execute code on startup of the repl - that's reintroduced
as a new parameter
 `--runBefore` now.